refactor: change package & Set-to-List realization
